requirements.txt
Django==1.7
MySQL-python==1.2.5
django-localflavor==1.0
wsgiref==0.1.2
